app/requirements.txt
asyncpg==0.22.0
autopep8==1.5.5
bcrypt==3.2.0
cffi==1.14.5
click==7.1.2
databases==0.4.1
DateTime==4.3
fastapi==0.63.0
uvloop
h11==0.12.0
passlib==1.7.4
psycopg2-binary==2.8.6
pycodestyle==2.6.0
pycparser==2.20
pydantic==1.7.3
pytz==2021.1
six==1.15.0
SQLAlchemy==1.3.23
starlette==0.13.6
toml==0.10.2
typing==3.7.4.3
uuid==1.30
uvicorn==0.13.4
zope.interface==5.2.0
passlib
gunicorn
python-json-logger
httptools